Remove obsolete remnants of the TPU testing infrastructure (TF-520)
This commit removes the remnants of the obsolete TPU testing
infrastructure. In the nearest future, we are planning to restore it
within the new tracing runtime (https://bugs.swift.org/browse/TF-520).
