+++ test/TensorFlowRuntime/conflicting_module.swift
@@ -11,11 +11,7 @@
 // REQUIRES: executable_test
 
 import StdlibUnittest
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 
 import ConflictingModule1
 import ConflictingModule2
