+++ test/TensorFlowRuntime/control_flow_1.swift
@@ -4,11 +4,7 @@
 // Control flow related tests.
 
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var ControlFlowTests = TestSuite("ControlFlow")
