+++ test/TensorFlowRuntime/control_flow_objc.swift
@@ -5,11 +5,7 @@
 // Control flow related tests.
 
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 import Foundation
