+++ test/TensorFlowRuntime/dataset_1.swift
@@ -1,23 +1,10 @@
 // RUN: %target-run-simple-swift %swift-tensorflow-test-run-extra-options
 // REQUIRES: executable_test
-//
-// Compiler-only testing for TPU graph lowering (e.g. shape requirements by XLA).
-
-// TODO: re-enable TPU compilation, when we are able to handle returning tensor
-// `three` below produced on TF CPU to the TPU graph function, without running
-// into the "missing tensor shape" issue. One option is to convert return
-// tensors to TF->host sends, so that in this case we can send directly from TF
-// CPU to host, even if the primary device is TPU.
 
 // Dataset tests.
 
 import TensorFlow
-
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var DatasetTests = TestSuite("Dataset")
