+++ test/TensorFlowRuntime/device_placement.swift
@@ -4,11 +4,7 @@
 // Test device placement API.
 
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var DevicePlacementTests = TestSuite("DevicePlacement")
