+++ test/TensorFlowRuntime/dynamic_compilation.swift
@@ -3,16 +3,12 @@
 
 import CTensorFlow
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var DynamicCompilationTests = TestSuite("DynamicCompilation")
 
-DynamicCompilationTests.testCPUOrGPU("ScalarNonConst") {
+DynamicCompilationTests.testAllBackends("ScalarNonConst") {
   _RuntimeConfig.printsDebugLog = true
   func scalarInitializer_CreateHostTensor(_ x: Float) {
     let y = Tensor<Float>(x)
@@ -21,7 +17,7 @@ DynamicCompilationTests.testCPUOrGPU("ScalarNonConst") {
   scalarInitializer_CreateHostTensor(1.2)
 }
 
-DynamicCompilationTests.testCPUOrGPU("AddFloat") {
+DynamicCompilationTests.testAllBackends("AddFloat") {
   _RuntimeConfig.printsDebugLog = true
   let x = Tensor<Float>(1.0)
   let y = Tensor<Float>(2.0)
@@ -29,7 +25,7 @@ DynamicCompilationTests.testCPUOrGPU("AddFloat") {
   expectNearlyEqualWithScalarTensor(3.0, z)
 }
 
-DynamicCompilationTests.testCPUOrGPU("AddInt64") {
+DynamicCompilationTests.testAllBackends("AddInt64") {
   _RuntimeConfig.printsDebugLog = true
   let x = Tensor<Int64>(1)
   let y = Tensor<Int64>(2)
@@ -37,7 +33,7 @@ DynamicCompilationTests.testCPUOrGPU("AddInt64") {
   expectEqualWithScalarTensor(3, z)
 }
 
-DynamicCompilationTests.testCPUOrGPU("AddInt32") {
+DynamicCompilationTests.testAllBackends("AddInt32") {
   _RuntimeConfig.printsDebugLog = true
   let x = Tensor<Int32>(1)
   let y = Tensor<Int32>(2)
@@ -45,7 +41,7 @@ DynamicCompilationTests.testCPUOrGPU("AddInt32") {
   expectEqualWithScalarTensor(3, z)
 }
 
-DynamicCompilationTests.testCPUOrGPU("2Adds") {
+DynamicCompilationTests.testAllBackends("2Adds") {
   _RuntimeConfig.printsDebugLog = true
   let x = Tensor<Float>(1.0)
   let y = Tensor<Float>(2.0)
