+++ test/TensorFlowRuntime/dynamic_compilation_tensor_group.swift
@@ -2,11 +2,7 @@
 // REQUIRES: executable_test
 
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var TensorGroupTest = TestSuite("TensorGroup")
