+++ test/TensorFlowRuntime/gpu_negative_tests.swift
@@ -5,11 +5,7 @@
 // This test suite contains expected failure tests for GPU execution.
 
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var GPUNegativeTests = TestSuite("GPUNegative")
