+++ test/TensorFlowRuntime/loops.swift
@@ -4,11 +4,7 @@
 // Loop tests.
 
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var LoopsTests = TestSuite("Loops")
