+++ test/TensorFlowRuntime/model_autodiff_runtime.swift
@@ -10,11 +10,7 @@
 
 import TensorFlow
 import StdlibUnittest
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 
 var ModelADTests = TestSuite("ModelAD")
