+++ test/TensorFlowRuntime/sese_loop_canonicalization.swift
@@ -2,11 +2,7 @@
 // REQUIRES: executable_test
 
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var ControlFlowTests = TestSuite("ControlFlow")
