+++ test/TensorFlowRuntime/tensor.swift
@@ -4,11 +4,7 @@
 // Tensor API tests.
 
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var TensorTests = TestSuite("Tensor")
@@ -186,6 +182,23 @@ TensorTests.testAllBackends("SliceIndexing") {
   expectEqual(Array(stride(from: 40.0, to: 60, by: 1)), array3D.scalars)
   expectEqual(Array(stride(from: 20.0, to: 30, by: 1)), array2D.scalars)
   expectEqual(Array(stride(from: 3.0, to: 5, by: 1)), array1D.scalars)
+}
+
+TensorTests.testAllBackends("SliceUpdate") {
+  var t1 = Tensor<Float>([[1, 2, 3], [4, 5, 6]])
+  t1[0] = Tensor(zeros: [3])
+  expectEqual(ShapedArray(shape:[2, 3], scalars: [0, 0, 0, 4, 5, 6]), t1.array)
+  var t2 = t1
+  t2[0][2] = Tensor(3)
+  expectEqual(ShapedArray(shape:[2, 3], scalars: [0, 0, 3, 4, 5, 6]), t2.array)
+  var t3 = Tensor<Bool>([[true, true, true], [false, false, false]])
+  t3[0][1] = Tensor(false)
+  expectEqual(ShapedArray(shape:[2, 3],
+                          scalars: [true, false, true, false, false, false]),
+              t3.array)
+  var t4 = Tensor<Bool>([[true, true, true], [false, false, false]])
+  t4[0] = Tensor(repeating: false, shape: [3])
+  expectEqual(ShapedArray(repeating: false, shape: [2, 3]), t4.array)
 }
 
 
@@ -399,8 +412,6 @@ TensorTests.testAllBackends("AdvancedIndexing") {
 }
 
 TensorTests.testAllBackends("Reduction") {
-  // TODO(b/111815968): triage and fix this TPU issue
-  #if !TPU
   // 2 x 5
   let x = Tensor<Float>([[1, 2, 3, 4, 5], [1, 2, 3, 4, 5]])
   expectEqual(Tensor(30), x.sum())
@@ -434,7 +445,6 @@ TensorTests.testAllBackends("Reduction") {
               x.variance(squeezingAxes: 1))
   expectEqual(Tensor(shape: [1, 2], scalars: [2, 2]),
               x.variance(alongAxes: 1))
-  #endif // !TPU
 }
 
 TensorTests.testAllBackends("Concatenation") {
@@ -681,6 +691,16 @@ TensorTests.testAllBackends("ReshapeTensor") {
   let y = Tensor<Float>(repeating: 0.0, shape: [1, 3, 1, 2, 1])
   let result = x.reshaped(like: y)
   expectEqual([1, 3, 1, 2, 1], result.shape)
+}
+
+TensorTests.testAllBackends("BroadcastTensor") {
+  // 1 -> 2 x 3 x 4
+  let one = Tensor<Float>(1)
+  var target = Tensor<Float>(repeating: 0.0, shape: [2, 3, 4])
+  let broadcasted = one.broadcast(like: target)
+  expectEqual(Tensor(repeating: 1, shape: [2, 3, 4]), broadcasted)
+  target .= Tensor(repeating: 1, shape: [1, 3, 1])
+  expectEqual(Tensor(repeating: 1, shape: [2, 3, 4]), target)
 }
 
 TensorTests.testAllBackends("Unbroadcast1") {
