+++ test/TensorFlowRuntime/tensor_autodiff_indirect.swift
@@ -5,11 +5,7 @@
 
 import TensorFlow
 import StdlibUnittest
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 
 var TensorADTests = TestSuite("TensorIndirectAD")
