+++ test/TensorFlowRuntime/tensor_autodiff_indirect_crasher.swift
@@ -8,11 +8,7 @@
 
 import TensorFlow
 import StdlibUnittest
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 
 var TensorADTests = TestSuite("TensorIndirectAD")
