+++ test/TensorFlowRuntime/tracer_tffunc.swift
@@ -4,11 +4,7 @@
 // Tracer tests.
 
 import TensorFlow
-#if TPU
-import TensorFlowUnittestTPU
-#else
 import TensorFlowUnittest
-#endif
 import StdlibUnittest
 
 var TracerTests = TestSuite("TracerTFFunction")
